+++ packages/jss-plugin-extend/src/index.test.js
@@ -132,6 +132,35 @@ describe('jss-plugin-extend', () => {
     })
   })
 
+  describe('multi mixed rule name and style objects extend', () => {
+    let sheet
+
+    beforeEach(() => {
+      const a = {float: 'left'}
+      sheet = jss.createStyleSheet({
+        b: {position: 'absolute'},
+        c: {
+          extend: [a, 'b'],
+          width: '1px'
+        }
+      })
+    })
+
+    it('should have correct output', () => {
+      expect(sheet.getRule('c')).to.not.be(undefined)
+      expect(sheet.toString()).to.be(
+        '.b-id {\n' +
+          '  position: absolute;\n' +
+          '}\n' +
+          '.c-id {\n' +
+          '  float: left;\n' +
+          '  position: absolute;\n' +
+          '  width: 1px;\n' +
+          '}'
+      )
+    })
+  })
+
   describe('nested extend 1', () => {
     let sheet