+++ changelog.md
@@ -4,6 +4,10 @@ Since you are interested in what happens next, in case, you work for a for-profi
 
 ---
 
+### Bug fixes
+
+- [jss-plugin-extend] Fix can not extend rule name is array [1357](https://github.com/cssinjs/jss/pull/1357)
+
 ### Improvements
 
 - [react-jss] add properly react default props types calculation (https://github.com/cssinjs/jss/pull/1353)